Refactor: 바텀시드 공통화
